Refactored git_integration. (#192)
The git_integration/ directory was getting problematic; everything was
being jammed in to git_repo.rs or git_wrap.rs, without much logic about
what was in each of those files. This PR simply moves the functionality
in those two files into logically defined source files and cleans up the
imports in this directory. It touches a lot, but there should be no
functionality change.

---------
