+++ rust/gitxet/tests/git_repo_xet_tests.rs
@@ -1,8 +1,8 @@
 #[cfg(test)]
 mod git_repo_tests_2 {
     use gitxetcore::errors::Result;
-    use gitxetcore::git_integration::git_repo::test_tools::TestRepo;
-    use gitxetcore::git_integration::git_repo::GitRepo;
+    use gitxetcore::git_integration::git_repo_test_tools::TestRepo;
+    use gitxetcore::git_integration::GitXetRepo;
     use std::path::Path;
 
     fn setup_path() {
@@ -12,7 +12,7 @@ mod git_repo_tests_2 {
         std::env::set_var("PATH", format!("{}:{}", path.to_str().unwrap(), &full_path));
     }
 
-    async fn verify_repo_structure(repo: &GitRepo) -> Result<(bool, bool)> {
+    async fn verify_repo_structure(repo: &GitXetRepo) -> Result<(bool, bool)> {
         let mut changed = false;
 
         changed |= repo.verify_or_create_xet_directories()?;
