+++ rust/gitxetcore/src/command.rs
@@ -38,8 +38,8 @@ use crate::config::{get_sanitized_invocation_command, ConfigGitPathOption};
 use crate::config_cmd::{handle_config_command, ConfigArgs};
 use crate::diff::{diff_command, DiffArgs};
 use crate::errors;
-use crate::git_integration::git_integration_plumb::{handle_hook_plumb_command, HookCommandShim};
-use crate::git_integration::git_wrap::perform_git_version_check;
+use crate::git_integration::git_version_checks::perform_git_version_check;
+use crate::git_integration::hook_command_entry::{handle_hook_plumb_command, HookCommandShim};
 use crate::log::{get_trace_span, initialize_tracing_subscriber};
 use crate::smudge_query_interface::SmudgeQueryPolicy;
 use crate::upgrade_checks::VersionCheckInfo;
