+++ rust/gitxetcore/src/command/filter.rs
@@ -5,7 +5,7 @@ use crate::config::XetConfig;
 use crate::constants;
 use crate::data_processing::PointerFileTranslator;
 use crate::errors;
-use crate::git_integration::git_repo::GitRepo;
+use crate::git_integration::GitXetRepo;
 use crate::stream::git_stream::GitStreamInterface;
 
 /// Filter command handler
@@ -14,7 +14,7 @@ pub async fn filter_command(config: XetConfig) -> errors::Result<()> {
     info!("Establishing Git Handshake.");
 
     // Sync up the notes to the local mdb
-    GitRepo::verify_repo_for_filter(config.clone()).await?;
+    GitXetRepo::verify_repo_for_filter(config.clone()).await?;
 
     let repo = PointerFileTranslator::from_config(&config).await?;
     let mut event_loop =
