+++ rust/gitxetcore/src/git_integration/mod.rs
@@ -1,7 +1,22 @@
 pub mod git_commits;
 pub mod git_file_tools;
-pub mod git_integration_plumb;
-pub mod git_notes_wrapper;
-pub mod git_repo;
+pub mod git_merkledb;
+mod git_notes_wrapper;
+mod git_process_wrapping;
+mod git_repo_paths;
+mod git_repo_plumbing;
+pub mod git_repo_salt;
+mod git_xet_repo;
+pub mod hook_command_entry;
+
 pub mod git_url;
-pub mod git_wrap;
+pub mod git_user_config;
+pub mod git_version_checks;
+
+pub use crate::git_integration::git_xet_repo::git_repo_test_tools; // HERE
+pub use git_file_tools::GitTreeListing;
+pub use git_notes_wrapper::GitNotesWrapper;
+pub use git_process_wrapping::*;
+pub use git_repo_paths::*;
+pub use git_repo_plumbing::*;
+pub use git_xet_repo::GitXetRepo;
